refactor(mobile): move things around in swap
