+++ apps/mobile/src/features/swap/swap-state/swap.constants.ts
